Added test files for classifier and validator

+++ include/core/fourdimensional_vector.h
@@ -11,22 +11,22 @@ public:
     /**
      * Creates a 4D vector given size of each vector and an initial value
      */
-    FourDimensional_Vector(size_t row, size_t col, size_t class_num, size_t shade_num, double default_value);
+    FourDimensional_Vector(size_t row, size_t col, size_t class_num, size_t shade_num, float default_value);
 
     /**
      * returns the value at the 4th level vector
      */
-    double GetValue(size_t row, size_t col, size_t class_label, size_t shade);
+    float GetValue(size_t row, size_t col, size_t class_label, size_t shade);
 
     /**
      * returns all of the values at the 4th level vector
      */
-    vector<double> GetShades(size_t row, size_t col, size_t class_label);
+    vector<float> GetShades(size_t row, size_t col, size_t class_label);
 
     /**
      * set the value at the 4th value vector given index at each vector
      */
-    void SetValue(size_t row, size_t col, size_t class_label, size_t shade, double val);
+    void SetValue(size_t row, size_t col, size_t class_label, size_t shade, float val);
     
     size_t GetRowSize();
 
@@ -37,7 +37,7 @@ public:
     size_t GetShadeSize();
 
 private:
-    vector<vector<vector<vector<double>>>> data_values_;
+    vector<vector<vector<vector<float>>>> data_values_;
     size_t row_size;
     size_t col_size;
     size_t class_size;
